msm: vidc_3x: increase timeout value for hw response and power collapse
When running concurrent video session, there are some possibilities
that CPU may take more than a second  to schedule hfi work handler.
In those cases, one of the threads which is waiting for the response
is timing-out.

Increase hw response timeout and power collapse timeout will give
more time for hfi work handler to be scheduled and process the response
messages.

CRs-Fixed: 1086284
Signed-off-by: Karthikeyan Periasamy <kperiasa@codeaurora.org>
Change-Id: I69eceb1530e154b53a0d6bd3db3b0e2b4b64dd67