Make msm_serial_hs RT to improve Bluetooth performance
msm_serial_hs threads for RX and TX are responsible for passing key
Bluetooth buffers, such as Bluetooth audio, between the Bluetooth HW and
the upper layers of the stack.  These threads are therefore on the
critical path for Bluetooth audio latency, and should be scheduled as
RT threads in order to ensure that they meet audio deadlines.

Cherry picked from:
commit e5ed4788fdcc6 ("Make msm_serial_hs RT to improve Bluetooth performance")

Bug: 36106419
Test: Play Bluetooth audio, confirm msm_serial_hs_0 threads are
      scheduled as RT threads via systrace
Signed-off-by: Philip Cuadra <philipcuadra@google.com>
Change-Id: I1ee92fb5c9ddfff4fe25d215ac635cad7688e40a
Signed-off-by: Petri Gynther <pgynther@google.com>