Subject: net: allow configuring default qdisc
Since 3.12 it has been possible to configure the default queuing
discipline via sysctl. This patch adds ability to configure the
default queue discipline in kernel configuration. This is useful for
environments where configuring the value from userspace is difficult
to manage.

The default is still the same as before (pfifo_fast) and it is
possible to change after kernel init with sysctl. This is similar
to how TCP congestion control works.

Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>
Change-Id: I7ce6e4784f69c243f54d990aa5eaf9cc358ef708
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Raphiel Rollerscaperers <rapherion@raphielgang.org>