kernel: Boost whenever a zygote-forked process becomes a top app
Boost to the max for 500 ms whenever the top app changes, which
improves app launch speeds and addresses jitter when switching between
apps. A check to make sure that the top-app's parent is zygote ensures
that a user-facing app is indeed what's added to the top app task group,
since app processes are forked from zygote.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: Iec1f69949fbd1cbfef4721e46e78e6a44e7ba2a7
Signed-off-by: Lau <laststandrighthere@gmail.com>