sched: reduce softirq conflicts with RT
This is a forward port of pa/890483 with modifications from the original
patch due to changes in sched/softirq.c which applies the same logic.

We're finding audio glitches caused by audio-producing RT tasks
that are either interrupted to handle softirq's or that are
scheduled onto cpu's that are handling softirq's.
In a previous patch, we attempted to catch many cases of the
latter problem, but it's clear that we are still losing
significant numbers of races in some apps.

This patch attempts to address both problems:
1. It prohibits handling softirq's when interrupting
   an RT task, by delaying the softirq to the ksoftirqd
   thread.
2. It attempts to reduce the most common windows in which
   we lose the race between scheduling an RT task on a remote
   core and starting to handle softirq's on that core.
   We still lose some races, but we lose significantly fewer.
   (And we don't want to introduce any heavyweight forms
   of synchronization on these paths.)

Bug: 64912585
Bug: 136771796
Signed-off-by: Miguel de Dios <migueldedios@google.com>
Change-Id: I8d14860b25b53c190f7bf3c1c9a78a6facad18b0
Signed-off-by: Albert I <kras@raphielgang.org>