UPSTREAM: wireguard: selftests: remove ancient kernel compatibility code
Quite a bit of the test suite was designed to work with ancient kernels.
Thankfully we no longer have to deal with this. This commit updates
things that we can finally update and removes things that we can finally
remove, to avoid the build-up of the last several years as a result of
having to support ancient kernels. We can finally rely on suppress_
prefixlength being available. On the build side of things, the no-PIE
hack is no longer required, and we can bump some of the tools, repair
our m68k and i686-kvm support, and get better coverage of the static
branches used in the crypto lib and in udp_tunnel.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: Iae3553274206b38be3c109028c74da24d062837c
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 9a69a4c8802adf642bc4a13d471b5a86b44ed434)
Bug: 152722841
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: Iae3553274206b38be3c109028c74da24d062837c
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
(cherry picked from commit 8dedc704d5a31baf7457969f4721178ef46c5c7b)
Signed-off-by: TogoFire <togofire@mailfence.com>