UPSTREAM: wireguard: noise: read preshared key while taking lock
Prior we read the preshared key after dropping the handshake lock, which
isn't an actual crypto issue if it races, but it's still not quite
correct. So copy that part of the state into a temporary like we do with
the rest of the handshake state variables. Then we can release the lock,
operate on the temporary, and zero it out at the end of the function. In
performance tests, the impact of this was entirely unnoticable, probably
because those bytes are coming from the same cacheline as other things
that are being copied out in the same manner.

Reported-by: Matt Dunwoodie <ncon@noconroy.net>
Fixes: e7096c131e51 ("net: WireGuard secure network tunnel")
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: I44a566ccb6648b66d9235a6aa2b2420e373bb099
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit bc67d371256f5c47d824e2eec51e46c8d62d022e)
Bug: 152722841
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: I44a566ccb6648b66d9235a6aa2b2420e373bb099
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
(cherry picked from commit c1737f415b330e475afeed8538b25e5fc523e478)
Signed-off-by: TogoFire <togofire@mailfence.com>