Makefile: Enable store motion pass for gcse
GCC enables global common subexpression elimination (-fgcse) with
-Os, -O2, -O3 optimisation levels. Enabling this flag runs a store
motion pass after gcse, which attempts to move the stores out of
loops. When used in conjunction with -fgcse-lm (load move, enabled
by default with -Os, -O2 and -O3 levels), loops with a load/store
sequence can be changed to a load before the loop and store after
the loop, hence improving loop code execution speed.

When enabled in conjunction with LTO and -fipa-pta, this yields a
marginally better hackbench score.
- Before: 47.6s
- After: 28.5s

Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Change-Id: Idcfe3c949e02838999fcae45cf7f42abb0797981
Signed-off-by: Jebaitedneko <Jebaitedneko@gmail.com>
Signed-off-by: TogoFire <italomellopereira@gmail.com>