leds:qpnp-flash-v2: Add support for flash realtime control
In the default design, every time we change the brightness
in torch node it gets reflected only when the enable bit
in the switch node is toggled from 0 to 1. Add new DT property
to optionally allow brightness change to reflect in realtime,
for as long as the switch node enable bit remains set.

Signed-off-by: Kavya Nunna <quic_knunna@quicinc.com>
Change-Id: Ib1d9d54d81fa7511bc9f492a59cb477a4d0927d8
Signed-off-by: TogoFire <togofire@mailfence.com>